requirements.txt
keras
tensorflow
matplotlib
pillow
numpy
requests